+++ src/java/com/idega/user/data/GroupDomainRelationBMPBean.java
@@ -88,7 +88,7 @@ public class GroupDomainRelationBMPBean extends GenericEntity implements GroupDo
   }
 
   public Integer getRelatedGroupPK(){
-    return (Integer)getIntegerColumnValue(RELATED_GROUP_ID_COLUMN);
+    return getIntegerColumnValue(RELATED_GROUP_ID_COLUMN);
   }
 
   public void setRelationship(GroupDomainRelationType type){
