+++ src/java/com/idega/util/Thumbnail.java
@@ -57,9 +57,9 @@ public  class Thumbnail
         int imageHeight = image.getHeight(null);
         double imageRatio = (double)imageWidth / imageHeight;
         if (thumbRatio < imageRatio)
-            thumbHeight = (int)((double)thumbWidth / imageRatio);
+            thumbHeight = (int)(thumbWidth / imageRatio);
         else
-            thumbWidth = (int)((double)thumbHeight * imageRatio);
+            thumbWidth = (int)(thumbHeight * imageRatio);
         BufferedImage thumbImage = new BufferedImage(thumbWidth, thumbHeight, 1);
         Graphics2D graphics2D = thumbImage.createGraphics();
         graphics2D.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_BILINEAR);
@@ -86,9 +86,9 @@ public  class Thumbnail
         int imageHeight = image.getHeight(null);
         double imageRatio = (double)imageWidth / imageHeight;
         if (thumbRatio < imageRatio)
-            thumbHeight = (int)((double)thumbWidth / imageRatio);
+            thumbHeight = (int)(thumbWidth / imageRatio);
         else
-            thumbWidth = (int)((double)thumbHeight * imageRatio);
+            thumbWidth = (int)(thumbHeight * imageRatio);
         BufferedImage thumbImage = new BufferedImage(thumbWidth, thumbHeight, 1);
         Graphics2D graphics2D = thumbImage.createGraphics();
         graphics2D.setRenderingHint(RenderingHints.KEY_INTERPOLATION, RenderingHints.VALUE_INTERPOLATION_BILINEAR);
